# Labels: optimise creation of signature with/without labels