--- a/promql/engine.go
+++ b/promql/engine.go
@@ -2051,7 +2051,7 @@ func (ev *evaluator) VectorBinop(op parser.ItemType, lhs, rhs Vector, matching *
 
 func signatureFunc(on bool, b []byte, names ...string) func(labels.Labels) string {
 	if on {
-	sort.Strings(names)
+		sort.Strings(names)
 		return func(lset labels.Labels) string {
 			return string(lset.BytesWithLabels(b, names...))
 		}
@@ -2094,15 +2094,7 @@ func resultMetric(lhs, rhs labels.Labels, op parser.ItemType, matching *parser.V
 
 	if matching.Card == parser.CardOneToOne {
 		if matching.On {
-		Outer:
-			for _, l := range lhs {
-				for _, n := range matching.MatchingLabels {
-					if l.Name == n {
-						continue Outer
-					}
-				}
-				enh.lb.Del(l.Name)
-			}
+			enh.lb.Keep(matching.MatchingLabels...)
 		} else {
 			enh.lb.Del(matching.MatchingLabels...)
 		}
@@ -2297,16 +2289,14 @@ func (ev *evaluator) aggregation(op parser.ItemType, grouping []string, without 
 		group, ok := result[groupingKey]
 		// Add a new group if it doesn't exist.
 		if !ok {
-			var m labels.Labels
-
+			lb.Reset(metric)
 			if without {
-				lb.Reset(metric)
 				lb.Del(grouping...)
 				lb.Del(labels.MetricName)
-				m = lb.Labels()
 			} else {
-				m = metric.WithLabels(grouping...)
+				lb.Keep(grouping...)
 			}
+			m := lb.Labels()
 			newAgg := &groupedAggregation{
 				labels:     m,
 				value:      s.V,